+++ SortText.c
@@ -594,7 +594,7 @@ int ft_sortFile(FTABLE* fp, int scope, char *fs, char *keys, char *sel, int sort
 	}
 
 	undo_startModification(fp);
-	if (find_setTextSelection(scope,wp,&mps,&mpe) == RNG_INVALID) {
+	if (find_setTextSelection(wp, scope, &mps,&mpe) == RNG_INVALID) {
 		return 0;
 	}
 	lpfirst = mps->lm;